Memory usage optimization just bcz im bored in school

+++ binding-mri/audio-binding.cpp
@@ -29,8 +29,8 @@
 	{ \
 		RB_UNUSED_PARAM; \
 		const char *filename; \
-		int volume = 100; \
-		int pitch = 100; \
+		unsigned short volume = 100; \
+		unsigned short pitch = 100; \
 		double pos = 0.0; \
 		if (rgssVer >= 3) \
 			rb_get_args(argc, argv, "z|iif", &filename, &volume, &pitch, &pos RB_ARG_END); \
@@ -56,8 +56,8 @@
 	{ \
 		RB_UNUSED_PARAM; \
 		const char *filename; \
-		int volume = 100; \
-		int pitch = 100; \
+		unsigned short volume = 100; \
+		unsigned short pitch = 100; \
 		rb_get_args(argc, argv, "z|ii", &filename, &volume, &pitch RB_ARG_END); \
 		GUARD_EXC( shState->audio().entity##Play(filename, volume, pitch); ) \
 		return Qnil; \
